requirements.txt
aiogram==3.13.1
openai>=1.30.0
aiosqlite>=0.20.0
apscheduler>=3.10.4
pytz>=2024.1
httpx>=0.27.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
pymupdf>=1.24.0
google-api-python-client>=2.130.0
google-auth>=2.29.0
